+++ CHANGELOG.md
@@ -29,6 +29,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 - feature/1201055143148093 - Add example outputs. Delete run.php
 - feature/1201235392821202 - Add type interfaces support for validators
 - feature/1201262042576522 - Add ArrayKeyValidator
+- feature/1201465645647982 - Add recreate validators from a JSON raw or dumped file
  
 ### Changed